+++ src/LightGraphs.jl
@@ -259,6 +259,7 @@ include("centrality13/katz.jl")
 include("centrality13/pagerank.jl")
 include("centrality13/eigenvector.jl")
 include("centrality13/radiality.jl")
+include("Community/Community.jl")
 include("community13/modularity.jl")
 include("community13/label_propagation.jl")
 include("community13/core-periphery.jl")
